extracted large chunk of parsing algorithm from Add function in Action.java and placed it in new function called processTaskInformation(Task userTask, String[] taskInformation)
It takes in a task object with a String[] containing the information of the task
as parsed by the main parser in determineTask(String userTaskDescription).

processTaskInformation returns a Task object with it's updated time and date
properties if it is a non-floating task.

This function can be used when reading from file, converting txt information into
the task objects.